requirement.txt
alembic==1.0.6
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
passlib==1.7.1
pbr==5.1.1
PyJWT==1.7.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
stevedore==1.30.0
virtualenv==16.2.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1